Remove assumption of there always being a displayedList
