+++ arches/app/src/components/ControlledListManager/ListTree.vue
@@ -1,13 +1,12 @@
 <script setup lang="ts">
-import arches from "arches";
-import { computed, ref } from "vue";
+import { inject, ref } from "vue";
 import { useGettext } from "vue3-gettext";
 
-import Button from "primevue/button";
-import Dropdown from "primevue/dropdown";
 import Tree from "primevue/tree";
 
 import LetterCircle from "@/components/ControlledListManager/LetterCircle.vue";
+import ListTreeControls from "@/components/ControlledListManager/ListTreeControls.vue";
+import { displayedRowKey, selectedLanguageKey } from "@/components/ControlledListManager/const.ts";
 import { bestLabel } from "@/components/ControlledListManager/utils.ts";
 
 import type { Ref } from "@/types/Ref";
@@ -18,114 +17,39 @@ import type {
     TreeSelectionKeys,
 } from "primevue/tree/Tree";
 
-import type { Language } from "@/types/arches";
-import type {
-    ControlledList,
-    ControlledListItem
-} from "@/types/ControlledListManager";
-
-const props: { displayedList: ControlledList } = defineProps(["displayedList"]);
-
-const selectedKey: Ref<typeof TreeSelectionKeys> = defineModel("selectedKey");
+const tree: Ref<typeof TreeNode[]> = ref([]);
+const selectedKeys: Ref<typeof TreeSelectionKeys> = ref({});
 const expandedKeys: Ref<typeof TreeExpandedKeys> = ref({});
-const selectedLanguage: Ref<Language> = defineModel("selectedLanguage");
+
+const { setDisplayedRow } = inject(displayedRowKey);
+const selectedLanguage = inject(selectedLanguageKey);
 
 const { $gettext } = useGettext();
 const lightGray = "#f4f4f4"; // todo: import from theme somewhere
 
-function itemAsNode(item: ControlledListItem): typeof TreeNode {
-    return {
-        key: item.id,
-        label: bestLabel(item, selectedLanguage.value.code).value,
-        children: item.children.map(child => itemAsNode(child)),
-        data: item,
-    };
-}
-
-function listAsNode(list: ControlledList): typeof TreeNode {
-    return {
-        key: list.id,
-        label: list.name,
-        children: list.items.map(item => itemAsNode(item)),
-        data: list,
-    };
-}
-
-const controlledListItemsTree = computed(() => {
-    return [listAsNode(props.displayedList)];
-});
-
-const itemClass = (id: string) => {
-    if (id in selectedKey.value) {
-        return "selected";
-    }
-    return "";
+const onRowSelect = (node: typeof TreeNode) => {
+    setDisplayedRow(node.data);
 };
-
-const expandAll = () => {
-    for (const node of controlledListItemsTree.value) {
-        expandNode(node);
-    }
-};
-
-const collapseAll = () => {
-    expandedKeys.value = {};
-};
-
-const expandNode = (node: typeof TreeNode) => {
-    if (node.children && node.children.length) {
-        expandedKeys.value[node.key] = true;
-
-        for (const child of node.children) {
-            expandNode(child);
-        }
-    }
-};
-
-// Runs on list switch if the parent remembers to send a :key
-expandAll();
 </script>
 
 <template>
-    <div class="controls">
-        <Button
-            class="control"
-            type="button"
-            icon="fa fa-plus"
-            :label="$gettext('Expand')"
-            @click="expandAll"
-        />
-        <Button
-            class="control"
-            type="button"
-            icon="fa fa-minus"
-            :label="$gettext('Collapse')"
-            @click="collapseAll"
-        />
-        <Dropdown
-            v-model="selectedLanguage"
-            :options="arches.languages"
-            option-label="name"
-            :placeholder="$gettext('Language')"
-            checkmark
-            :highlight-on-select="false"
-            :pt="{
-                root: { class: 'control' },
-                input: { style: { fontFamily: 'inherit', fontSize: 'small', textAlign: 'center' } },
-                itemLabel: { style: { fontSize: 'small' } },
-            }"
-        />
-    </div>
+    <ListTreeControls
+        v-model="tree"
+        v-model:expanded-keys="expandedKeys"
+        v-model:selected-keys="selectedKeys"
+        :selected-keys
+    />
     <Tree
-        v-model:selectionKeys="selectedKey"
-        :value="controlledListItemsTree"
+        v-if="tree"
+        v-model:selectionKeys="selectedKeys"
+        :value="tree"
         :expanded-keys
         :filter="true"
         filter-mode="lenient"
         :filter-placeholder="$gettext('Find')"
-        selection-mode="single"
+        selection-mode="checkbox"
         :pt="{
-            root: { style: { flexGrow: 1 } },
+            root: { style: { flexGrow: 1, margin: '1rem' } },
             input: {
                 style: { height: '3.5rem', fontSize: '14px' },
             },
@@ -138,18 +62,18 @@ expandAll();
             }),
             label: { style: { textWrap: 'nowrap', marginLeft: '0.5rem' } },
         }"
+        @node-select="onRowSelect"
     >
         <template #nodeicon="slotProps">
             <LetterCircle :labelled="slotProps.node.data" />
         </template>
         <template #default="slotProps">
-            {{ slotProps.node.label }}
+            {{ slotProps.node.data.name ?? bestLabel(slotProps.node.data, selectedLanguage.code).value }}
             <span v-if="slotProps.node.data.uri">
                 (<a
                     :href="slotProps.node.data.uri"
                     target="_blank"
                     rel="noopener noreferrer"
-                    :class="itemClass(slotProps.node.data.id)"
                 >{{ slotProps.node.data.uri }}</a>)
             </span>
         </template>
@@ -159,26 +83,5 @@ expandAll();
 <style scoped>
 a {
     color: var(--blue-500);
-}
-.controls {
-    display: flex;
-    background: #f3fbfd;
-    padding: 1rem;
-    gap: 0.5rem;
-}
-.control {
-    flex: 0.33;
-    border: 0;
-    background: lightgray;
-    font-size: small;
-}
-.button {
-    font-size: small;
-    height: 4rem;
-    margin: 0.5rem;
-    justify-content: center;
-    font-weight: 600;
-    color: white;
-    text-wrap: nowrap;
 }
 </style>
