+++ arches/app/src/components/ControlledListManager/const.ts
@@ -1,11 +1,14 @@
-// Injection keys
 import type { ControlledList, ControlledListItem } from "@/types/ControlledListManager";
+import type { Language } from "@/types/arches";
+
+// Injection keys
 // Cannot import this from vue? 🤔
 // eslint-disable-next-line @typescript-eslint/no-unused-vars, @typescript-eslint/ban-types
 interface InjectionKey<T> extends Symbol {}
 
-export const displayedListKey = Symbol() as InjectionKey<ControlledList | null>;
+export const displayedRowKey = Symbol() as InjectionKey<ControlledList | null>;
 export const itemKey = Symbol() as InjectionKey<ControlledListItem | null>;
+export const selectedLanguageKey = Symbol() as InjectionKey<Language | null>;
 
 // Constants
 export const PREF_LABEL = "prefLabel";
